+++ aiida/cmdline/commands/cmd_computer.py
@@ -12,7 +12,7 @@
 import sys
 import click
 
-from aiida.cmdline.commands import verdi
+from aiida.cmdline.commands.cmd_verdi import verdi
 from aiida.cmdline.params import options, arguments
 from aiida.cmdline.params import types
 from aiida.cmdline.params.options.interactive import InteractiveOption
