+++ aiida/cmdline/commands/cmd_daemon.py
@@ -16,7 +16,7 @@ import time
 import click
 from click_spinner import spinner
 
-from aiida.cmdline.commands import verdi
+from aiida.cmdline.commands.cmd_verdi import verdi
 from aiida.cmdline.utils import decorators
 from aiida.cmdline.utils.common import get_env_with_venv_bin
 from aiida.cmdline.utils.daemon import get_daemon_status, print_client_response_status
