+++ aiida/cmdline/commands/cmd_export.py
@@ -11,7 +11,7 @@
 """`verdi export` command."""
 import click
 
-from aiida.cmdline.commands import verdi
+from aiida.cmdline.commands.cmd_verdi import verdi
 from aiida.cmdline.params import arguments
 from aiida.cmdline.params import options
 from aiida.cmdline.utils import echo
