+++ aiida/cmdline/commands/cmd_import.py
@@ -10,7 +10,7 @@
 """`verdi import` command."""
 import click
 
-from aiida.cmdline.commands import verdi
+from aiida.cmdline.commands.cmd_verdi import verdi
 from aiida.cmdline.params.options import MultipleValueOption
 from aiida.cmdline.utils import decorators, echo
