+++ aiida/cmdline/commands/cmd_profile.py
@@ -12,7 +12,7 @@ This allows to manage profiles from command line.
 """
 import click
 
-from aiida.cmdline.commands import verdi
+from aiida.cmdline.commands.cmd_verdi import verdi
 from aiida.cmdline.utils import echo
 from aiida.cmdline.params import options
 from aiida.control.postgres import Postgres
