+++ aiida/cmdline/commands/cmd_rehash.py
@@ -10,7 +10,7 @@
 """`verdi rehash` command."""
 import click
 
-from aiida.cmdline.commands import verdi
+from aiida.cmdline.commands.cmd_verdi import verdi
 from aiida.cmdline.params import arguments
 from aiida.cmdline.params.types.plugin import PluginParamType
 from aiida.cmdline.utils import decorators, echo
