+++ aiida/cmdline/commands/cmd_run.py
@@ -6,7 +6,7 @@ import sys
 
 import click
 
-from aiida.cmdline.commands import verdi
+from aiida.cmdline.commands.cmd_verdi import verdi
 from aiida.cmdline.params.options.multivalue import MultipleValueOption
 from aiida.cmdline.utils import decorators, echo
