+++ aiida/cmdline/commands/cmd_setup.py
@@ -9,7 +9,7 @@
 # For further information please visit http://www.aiida.net               #
 ###########################################################################
 """`verdi setup` command."""
-from aiida.cmdline.commands import verdi
+from aiida.cmdline.commands.cmd_verdi import verdi
 from aiida.cmdline.params import arguments, options
 from aiida.control.profile import setup_profile
