+++ aiida/cmdline/commands/cmd_shell.py
@@ -13,7 +13,7 @@ from __future__ import absolute_import
 import os
 import click
 
-from aiida.cmdline.commands import verdi
+from aiida.cmdline.commands.cmd_verdi import verdi
 from aiida.cmdline.utils import decorators, echo
